+++ templates/react-native/shared/Shared.js
@@ -6,22 +6,32 @@ import {
   Image,
   Text,
   View,
+  ViewStyle,
+  TextStyle,
 } from 'react-native';
 
-import type {
-  ____ViewStyleProp_Internal as ViewStyle,
-  ____TextStyleProp_Internal as TextStyle,
-  ____ImageStyleProp_Internal as ImageStyle,
-} from 'react-native/Libraries/StyleSheet/StyleSheetTypes';
+// import type {
+//   ____ViewStyleProp_Internal as ViewStyle,
+//   ____TextStyleProp_Internal as TextStyle,
+//   ____ImageStyleProp_Internal as ImageStyle,
+// } from 'react-native/Libraries/StyleSheet/StyleSheetTypes';
 
 import { ratio, colors } from '../../utils/Styles';
+
+type Styles = {
+  wrapper: ViewStyle,
+  text: TextStyle,
+};
 
 const styles = StyleSheet.create({
   wrapper: {
     backgroundColor: 'transparent',
-
     alignItems: 'center',
     justifyContent: 'center',
+  },
+  text: {
+    fontSize: 20,
+    color: 'black',
   },
 });
 
@@ -41,7 +51,7 @@ class Shared extends Component<Props, State> {
   render() {
     return (
       <View style={this.props.style}>
-        <Text>Shared</Text>
+        <Text style={styles.textStyle}>Shared</Text>
       </View>
     );
   }