+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelogs
 **[1.2+]**
 * Get rid of `path` and `fs` as in [#5](https://github.com/dooboolab/dooboo-cli/issues/4);
+* Enhanced template for styling style props.
 **[1.1+]**
 * React Native template added `react-native-gesture-handler` to be linked.
 * Removed all repositories to `dooboolab` repo.